Polish "Auto-configure Jetty connection and SSL metrics"
See gh-26418
