+++ spring-boot-project/spring-boot-actuator/src/main/java/org/springframework/boot/actuate/metrics/web/jetty/JettyServerThreadPoolMetricsBinder.java
@@ -21,23 +21,16 @@ import java.util.Collections;
 import io.micrometer.core.instrument.MeterRegistry;
 import io.micrometer.core.instrument.Tag;
 import io.micrometer.core.instrument.binder.jetty.JettyServerThreadPoolMetrics;
+import org.eclipse.jetty.server.Server;
 import org.eclipse.jetty.util.thread.ThreadPool;
 
-import org.springframework.boot.context.event.ApplicationStartedEvent;
-import org.springframework.boot.web.context.WebServerApplicationContext;
-import org.springframework.boot.web.embedded.jetty.JettyWebServer;
-import org.springframework.boot.web.server.WebServer;
-import org.springframework.context.ApplicationContext;
-import org.springframework.context.ApplicationListener;
-
 /**
- * Binds {@link JettyServerThreadPoolMetrics} in response to the
- * {@link ApplicationStartedEvent}.
+ * {@link AbstractJettyMetricsBinder} for {@link JettyServerThreadPoolMetrics}.
  *
  * @author Andy Wilkinson
  * @since 2.1.0
  */
-public class JettyServerThreadPoolMetricsBinder implements ApplicationListener<ApplicationStartedEvent> {
+public class JettyServerThreadPoolMetricsBinder extends AbstractJettyMetricsBinder {
 
 	private final MeterRegistry meterRegistry;
 
@@ -53,22 +46,11 @@ public class JettyServerThreadPoolMetricsBinder implements ApplicationListener<A
 	}
 
 	@Override
-	public void onApplicationEvent(ApplicationStartedEvent event) {
-		ApplicationContext applicationContext = event.getApplicationContext();
-		ThreadPool threadPool = findThreadPool(applicationContext);
+	protected void bindMetrics(Server server) {
+		ThreadPool threadPool = server.getThreadPool();
 		if (threadPool != null) {
 			new JettyServerThreadPoolMetrics(threadPool, this.tags).bindTo(this.meterRegistry);
 		}
-	}
-
-	private ThreadPool findThreadPool(ApplicationContext applicationContext) {
-		if (applicationContext instanceof WebServerApplicationContext) {
-			WebServer webServer = ((WebServerApplicationContext) applicationContext).getWebServer();
-			if (webServer instanceof JettyWebServer) {
-				return ((JettyWebServer) webServer).getServer().getThreadPool();
-			}
-		}
-		return null;
 	}
 
 }
